use main.py as pipeline runner and bkup main_old
